refactor(dtm): inline group module into corpus
Closes #297
