+++ penelope/corpus/dtm/__init__.py
@@ -1,6 +1,5 @@
 # type: ignore
 
 from .corpus import VectorizedCorpus, find_matching_words_in_vocabulary
-from .group import GroupByMixIn
 from .interface import IVectorizedCorpus
 from .store import StoreMixIn, load_metadata, store_metadata
