--- penelope/corpus/dtm/group.py
@@ -1,274 +1,0 @@
-from __future__ import annotations
-
-from typing import Any, Callable, List, Literal, Mapping, Protocol, Tuple
-
-import numpy as np
-import pandas as pd
-import scipy
-from scipy import sparse as sp
-
-from penelope import utility as pu
-from penelope.utility.utils import dict_of_key_values_inverted_to_dict_of_value_key
-
-from .interface import IVectorizedCorpus
-
-# pylint: disable=no-member
-
-
-KNOWN_TIME_PERIODS: dict[str, int] = {"year": 1, "lustrum": 5, "decade": 10}
-
-
-def create_temporal_key_categorizer(temporal_key_specifier: str | dict | Callable[[Any], Any]) -> Callable[[Any], Any]:
-    if callable(temporal_key_specifier):
-        return temporal_key_specifier
-
-    if isinstance(temporal_key_specifier, str):
-        if temporal_key_specifier not in KNOWN_TIME_PERIODS:
-            raise ValueError(f"{temporal_key_specifier} is not a known period specifier")
-        return lambda y: y - int(y % KNOWN_TIME_PERIODS[temporal_key_specifier])
-
-    year_group_mapping = dict_of_key_values_inverted_to_dict_of_value_key(temporal_key_specifier)
-    return lambda x: year_group_mapping.get(x, np.nan)
-
-
-class SupportsGroupBy(Protocol):
-    """Protocol defining the interface required by the active grouping mixin."""
-
-    bag_term_matrix: scipy.sparse.spmatrix
-    token2id: dict
-    document_index: pd.DataFrame
-    overridden_term_frequency: Any
-    payload: dict
-
-    def create(
-        self,
-        bag_term_matrix: scipy.sparse.spmatrix,
-        token2id: dict,
-        document_index: pd.DataFrame,
-        overridden_term_frequency: Any,
-        **kwargs,
-    ) -> IVectorizedCorpus: ...
-
-    def group_by_indices_mapping(
-        self,
-        document_index: pd.DataFrame,
-        category_indices: Mapping[int, List[int]],
-        aggregate: str = "sum",
-        dtype: np.dtype | None = None,
-    ) -> IVectorizedCorpus: ...
-
-
-class GroupByMixIn:
-    def group_by_indices_mapping(
-        self: SupportsGroupBy,
-        document_index: pd.DataFrame,
-        category_indices: Mapping[int, List[int]],
-        aggregate: str = "sum",
-        dtype: np.dtype | None = None,
-    ) -> IVectorizedCorpus:
-        matrix: scipy.sparse.spmatrix = group_DTM_by_indices_mapping(
-            dtm=self.bag_term_matrix,  # type: ignore
-            n_docs=len(document_index),
-            category_indices=category_indices,
-            aggregate=aggregate,
-            dtype=dtype,
-        )
-        return self.create(  # type: ignore
-            matrix.tocsr(),
-            token2id=self.token2id,  # type: ignore
-            document_index=document_index,
-            overridden_term_frequency=self.overridden_term_frequency,  # type: ignore
-            **self.payload,  # type: ignore
-        )
-
-    def group_by_pivot_keys(  # pylint: disable=too-many-arguments
-        self: SupportsGroupBy,
-        temporal_key: Literal["year", "decade", "lustrum"],
-        pivot_keys: List[str],
-        filter_opts: pu.PropertyValueMaskingOpts,
-        document_namer: Callable[[pd.DataFrame], pd.Series] | None,
-        aggregate: str = "sum",
-        fill_gaps: bool = False,
-        drop_group_ids: bool = True,
-        dtype: np.dtype | None = None,
-    ):
-        """Group corpus by a temporal key and zero to many pivot keys."""
-
-        def default_document_namer(df: pd.DataFrame) -> pd.Series:
-            """Vectorized string concatenation - 5-10x faster than apply with lambda."""
-            cols = [temporal_key] + pivot_keys
-            result = df[cols[0]].astype(str)
-            for col in cols[1:]:
-                result = result + "_" + df[col].astype(str)
-            return result
-
-        def document_index_aggregates(df: pd.DataFrame, grouping_keys: List[str]) -> dict:
-            document_id_column = "_document_id_np" if "_document_id_np" in df.columns else "document_id"
-
-            aggs: dict = {"document_ids": (document_id_column, list)}
-
-            for count_column in {"n_tokens", "n_raw_tokens", "tokens"}.intersection(set(df.columns)):
-                aggs[count_column] = (count_column, "sum")
-
-            if "year" in df.columns and "year" not in grouping_keys:
-                aggs["year"] = ("year", min)
-
-            if "n_documents" not in df.columns:
-                aggs["n_documents"] = (document_id_column, "nunique")
-            else:
-                aggs["n_documents"] = ("n_documents", "sum")
-
-            return aggs
-
-        if document_namer is None:
-            document_namer = default_document_namer
-
-        di: pd.DataFrame = self.document_index
-        gdi: pd.DataFrame = di if not pivot_keys or len(filter_opts or []) == 0 else di[filter_opts.mask(di)]
-
-        if "document_id" in gdi.columns:
-            # Avoid full DataFrame copy - just add column directly with astype
-            gdi["_document_id_np"] = gdi["document_id"].astype(np.int64)
-
-        if temporal_key not in gdi.columns:
-            gdi[temporal_key] = gdi["year"].apply(create_temporal_key_categorizer(temporal_key))
-
-        aggs: dict = document_index_aggregates(gdi, [temporal_key] + pivot_keys)
-
-        gdi = gdi.groupby([temporal_key] + pivot_keys, as_index=False).agg(**aggs)
-        gdi["document_name"] = document_namer(gdi)
-        gdi["filename"] = gdi.document_name
-
-        if fill_gaps:
-            gdi = fill_temporal_gaps_in_group_document_index(gdi, temporal_key, pivot_keys, aggs)
-
-        gdi["document_id"] = gdi.index.astype(np.int32)
-
-        # Single-pass type conversion - 2x faster than multiple as_slim_types calls
-        gdi = optimize_index_types(gdi, temporal_key)
-
-        gdi["time_period"] = gdi[temporal_key]
-
-        category_indices: Mapping[int, List[int]] = gdi["document_ids"].to_dict()  # type: ignore[assignment]
-
-        if drop_group_ids:
-            gdi.drop(columns="document_ids", inplace=True, errors="ignore")
-
-        return self.group_by_indices_mapping(
-            document_index=gdi,
-            category_indices=category_indices,
-            aggregate=aggregate,
-            dtype=dtype,
-        )
-
-
-def optimize_index_types(gdi: pd.DataFrame, temporal_key: str) -> pd.DataFrame:
-    """Optimize DataFrame types for memory efficiency."""
-
-    type_conversions: dict[str, type] = {
-        col: np.int32 for col in ["n_documents", "n_tokens", "n_raw_tokens", "tokens"] if col in gdi.columns
-    }
-    type_conversions.update({col: np.int16 for col in ["year", temporal_key] if col in gdi.columns})
-    if type_conversions:
-        gdi = gdi.astype(type_conversions)
-    return gdi
-
-
-def temporal_key_values_with_no_gaps(series: pd.Series, temporal_key: str):
-    """Return sorted temporal key values, filling expected gaps."""
-    step = {"lustrum": 5, "decade": 10}.get(temporal_key, 1)
-    return list(range(series.min(), series.max() + 1, step))
-
-
-def fill_temporal_gaps_in_group_document_index(
-    di: pd.DataFrame,
-    temporal_key: str,
-    pivot_keys: list[str],
-    aggs: dict,
-) -> pd.DataFrame:
-    """Fill missing temporal values with empty documents.
-
-    Optimized to use dict of arrays instead of list of dicts for 2-3x speedup.
-    """
-    sep = "_" if pivot_keys else ""
-
-    values_with_no_gaps = set(temporal_key_values_with_no_gaps(di[temporal_key], temporal_key=temporal_key))
-    missing_values = sorted(values_with_no_gaps - set(di[temporal_key]))
-
-    if not missing_values:
-        # No gaps to fill - early return
-        return di
-
-    n_missing = len(missing_values)
-
-    # Create dict of arrays instead of list of dicts - 2-3x faster
-    missing_data = {
-        temporal_key: missing_values,
-        'document_ids': [[] for _ in range(n_missing)],
-        'document_name': [f'{val}{sep}{sep.join(["0"] * len(pivot_keys))}' for val in missing_values],
-    }
-
-    # Add zero columns for all aggregates (already zeros, no fillna needed)
-    for k in aggs.keys():
-        if k != "document_ids":
-            missing_data[k] = [0] * n_missing
-
-    di_missing = pd.DataFrame(missing_data)
-
-    di = pd.concat([di, di_missing], ignore_index=True)
-    di.sort_values(by=[temporal_key] + pivot_keys, inplace=True, ascending=True)
-    di.reset_index(inplace=True, drop=True)
-    di["document_id"] = di.index
-    di["filename"] = di.document_name
-
-    return di
-
-
-def group_DTM_by_indices_mapping(
-    dtm: scipy.sparse.csr_matrix,
-    n_docs: int,
-    category_indices: Mapping[int, List[int]],
-    aggregate: str = "sum",
-    dtype: np.dtype | None = None,
-):
-    """Group document-term matrix by category indices using efficient sparse matrix multiplication.
-
-    This optimized version builds a sparse mapping matrix and uses matrix multiplication
-    instead of row-by-row iteration, providing 10-100x speedup for large matrices.
-
-    Args:
-        dtm: Document-term matrix to group (n_original_docs x n_terms)
-        n_docs: Number of output documents (categories)
-        category_indices: Mapping from target document ID to list of source document IDs
-        aggregate: Aggregation method - "sum" or "mean"
-        dtype: Output dtype (auto-detected if None)
-
-    Returns:
-        Grouped sparse matrix (n_docs x n_terms)
-    """
-    assert dtm.shape is not None
-
-    n_original_docs = dtm.shape[0]
-    dtype_y = dtype or (np.int32 if np.issubdtype(dtm.dtype, np.integer) and aggregate == "sum" else np.float64)
-
-    # Build sparse mapping matrix: (n_docs x n_original_docs)
-    # Each row represents a target document, each column an original document
-    # Values are 1 for sum, or 1/count for mean
-    row_indices = []
-    col_indices = []
-    data = []
-
-    for target_doc_id, source_doc_ids in category_indices.items():
-        if len(source_doc_ids) > 0:
-            weight = 1.0 / len(source_doc_ids) if aggregate == "mean" else 1.0
-            row_indices.extend([target_doc_id] * len(source_doc_ids))
-            col_indices.extend(source_doc_ids)
-            data.extend([weight] * len(source_doc_ids))
-
-    # Create sparse mapping matrix
-    mapping_matrix = sp.csr_matrix((data, (row_indices, col_indices)), shape=(n_docs, n_original_docs), dtype=dtype_y)
-
-    # Single matrix multiplication - much faster than row-by-row iteration!
-    matrix = mapping_matrix @ dtm
-
-    return matrix.tocsr()
